Updating file structure
