--- doc/_config.yml
@@ -1,3 +1,0 @@
-theme: jekyll-theme-minimal
-future: true
-title: Welcome to FRIDGe
