+++ docs/source/ka_br.md
@@ -1,0 +1,4 @@
+# Knowledge Agents - Blackboard Reader (KA-BR)
+
+KA-BR agents are used to move solutions on the blackboard.
+There are three agents required for use in
